LayoutTests/js/dom/domjit-function-get-element-by-id-licm.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<div id="parentNode">
<div id="previousSibling"></div><div id="target"><div id="firstChild"></div><div id="lastChild"></div></div><div id="nextSibling"></div>
</div>

<script>
description('Test DOMJIT function getElementById will be LICM-ed.');

function test() {
    var ret = 0;
    for (var i = 0; i < 1e4; ++i) {
        ret = document.getElementById("target").nodeType;
        ret = document.getElementById("target").nodeType;
        ret = document.getElementById("target").nodeType;
        ret = document.getElementById("target").nodeType;
    }
    return ret;
}
var result;
(function () {
    for (var i = 0; i < 100; ++i) {
        result = test();
        shouldBe(`result`, `1`, true);
    }
}());
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
